% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{count_alleles}
\alias{count_alleles}
\title{count_alleles}
\usage{
count_alleles(GT)
}
\arguments{
\item{GT}{a matrix of genotypes represented by strings}
}
\value{
A DataFramw.
}
\description{
Summarize genotypes and alleles.
}
\details{
The function \code{rgt} generates a matrix of random genotypes.
}
