% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneticDiff.R
\name{GeneticDiff}
\alias{GeneticDiff}
\title{GeneticDiff}
\description{
Genetic differentiation
}
\details{
GeneticDiff provides tools to explore and measure genetic differentiation.
Functionality is included to simulate data, measure within population diversity and among population differentiation.
}
