% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{genotype_split}
\alias{genotype_split}
\title{genotype_split - split a genotype into alleles}
\usage{
genotype_split(myGT)
}
\arguments{
\item{myGT}{a string containing a genotype}
}
\description{
Split a genotype into alleles based on the delimiters forward slash (/)  and pipe (|).
}
\details{
The funciton \code{genotype_split} splits a genotype into a vector of alleles.
The alleles are expected to be delimited by either a forward slash ('/') or a pipe ('|').
This format is expected to conform with the VCF specification but may handle other formats as well.
}
\examples{
genotype_split("0/1/1|2")


}
