% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{rDNA}
\alias{rDNA}
\title{rDNA - generate random DNA sequences}
\usage{
rDNA(length = 100, prob = c(0.25, 0.25, 0.25, 0.25))
}
\arguments{
\item{length}{the length, in nucleotide positions, of the sequence to simulate}

\item{prob}{a vector of weights for the four possible nucleotides (A, C, G, and T)}
}
\value{
A matrix of nucleotides consisting of one row (sample) that should be compatible with \code{ape::as.DNAbin()}.
}
\description{
generate random DNA sequences
}
\details{
Uses the function \code{base::sample()} to generate random sequences of nucleotides.
}
\examples{
myDNA <- rDNA()

\dontrun{
ape::as.DNAbin(myDNA)
}

}
\seealso{
\code{base::sample()}
\code{ape::as.DNAbin()}
}
