% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rgt}
\alias{rgt}
\title{rgt}
\usage{
rgt(nsamp = 4L, nvar = 3L, pphased = 50L, pploid = as.integer(c(0,
  100)), pallele = as.integer(c(50, 50)))
}
\arguments{
\item{nsamp}{number of samples to simulate}

\item{nvar}{number of variants to simulate}

\item{pphased}{probability each genotype will be phased}

\item{pploid}{probability each genotype will be a ploidy level}

\item{pallele}{probability of each allele}
}
\description{
Generate a matrix of random genotypes.
}
\details{
Generate a matrix of random genotypes.
}
\examples{
rgt()


}
