% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rgt}
\alias{rgt}
\title{rgt}
\usage{
rgt(nsamp = 4L, nvar = 3L, pphased = as.numeric(c(0.5)),
  pploid = as.numeric(c(0, 1)), pallele = as.numeric(c(0.5, 0.5)))
}
\arguments{
\item{nsamp}{number of samples to simulate}

\item{nvar}{number of variants to simulate}

\item{pphased}{probability each genotype will be phased}

\item{pploid}{probability each genotype will be a ploidy level}

\item{pallele}{probability of each allele}
}
\description{
Generate a matrix of random genotypes.
}
\details{
Generate a matrix of random genotypes.
}
